refactor(theme): Remove seekUnifiedBeta base tokens
re #513
